requirements.txt
TikTokApi
playwright
python-dotenv
matplotlib
networkx
scipy
flask
mpld3